requirements.txt
apipkg==1.5
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
click==7.1.1
execnet==1.7.1
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.1
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.1
pytest-forked==1.1.3
python-dotenv==0.12.0
six==1.14.0
SQLAlchemy==1.3.15
SQLAlchemy-Utils==0.36.3
typed-ast==1.4.1
wcwidth==0.1.9
Werkzeug==1.0.0
wrapt==1.11.2
zipp==3.1.0
